+++ inst/scripts/germany_example.R
@@ -4,8 +4,14 @@ library(data.table)
 
 # Load and filter germany hospitalisations
 nat_germany_hosp <- germany_covid19_hosp[location == "DE"][age_group %in% "00+"]
-nat_germany_hosp <- nat_germany_hosp[report_date <= as.Date("2021-10-01")]
+nat_germany_hosp <- enw_filter_report_dates(
+  nat_germany_hosp, latest_date = "2021-10-01"
+)
 
+# Make sure observations are complete
+nat_germany_hosp <- enw_complete_dates(
+  nat_germany_hosp, by = c("location", "age_group")
+)
 # Make a retrospective dataset
 retro_nat_germany <- enw_filter_report_dates(
   nat_germany_hosp, remove_days = 40
@@ -17,7 +23,7 @@ retro_nat_germany <- enw_filter_reference_dates(
 # Get latest observations for the same time period
 latest_obs <- enw_latest_data(nat_germany_hosp)
 latest_obs <- enw_filter_reference_dates(
-  latest_obs, remove_days = 40, include_days = 40
+  latest_obs, remove_days = 40, include_days = 20
 )
 
 # Preprocess observations